plant disease detection/requirement.txt
Flask
numpy
Keras==2.2.4
gevent
pillow
h5py
tensorflow==1.5.0 
opencv-python
